pyproject.toml
[project]
name = "asr-tools"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "datasets[audio]>=3.3.2",
    "huggingface-hub[cli,hf-transfer]>=0.29.3",
    "jiwer>=3.1.0",
    "librosa>=0.11.0",
    "loguru>=0.7.3",
    "loguru-config>=0.1.0",
    "nemo-toolkit[asr]>=2.1.0",
    "openai-whisper>=20240930",
    "polars>=1.25.2",
    "pyyaml>=6.0.2",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "transformers[torch]>=4.49.0",
]
